+++ frontend/src/app/App.js
@@ -5,7 +5,7 @@ import Header from './components/header';
 import {Container} from './../styles/blocks/layout';
 // pages 
 import Dashboard from './pages/Dashboard';
-import Match from './pages/Match';
+import CreateMatch from './pages/CreateMatch';
 
 
 class App extends Component {
@@ -16,7 +16,7 @@ class App extends Component {
           <Header />
           
           <Route exact path={ROUTES.DASHBOARD} component={Dashboard} />
-          <Route exact path={ROUTES.MATCH} component={Match} />
+          <Route exact path={ROUTES.CREATEMATCH} component={CreateMatch} />
 
         </Container>
       </Router>
