+++ frontend/src/app/components/SelectInput/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { SelectBox } from './../../../styles/blocks';
+
+const SelectInput = ({ data }) => (
+  <SelectBox>
+    {data.map((data) => 
+      <option value={data.name}>
+        {data.name}
+      </option>  
+    )}
+  </SelectBox>
+)
+
+export default SelectInput;
+
+
